requirements.txt
Django
gunicorn
psycopg2-binary
redis
python-geoip
python-geoip-geolite2
Pillow
PyInstaller
pyscreenshot
django-crispy-forms
django-ipware
